backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
beautifulsoup4
playwright
Pillow
openai
supabase
daytona-sdk
python-multipart
